binary search tree tests
fixed a bug in it too. preOrder traversal was broken

also added chaining for applicable methods

Level order traversal was completely wrong. made it work and cleaned it
up

which included me adding a length parameter to my quene data structure
to help with the level traversal.

You could use an array but I had tested my quene so figured why not use
it