min heap with tests
also:
 - needed to make sure my sorter dropped values less than -1 in siftDown
   since negative ones will end up at the end of the array because they
   are the lowest values

P.S.

man javascript inheritance can get quirky

had a really wierd issue in the tests where the comparator was getting
over written in the different functions.

Solution? make sure to use a different variable name for them and not
try to use the same one. Thus the maxHeap and minHeap everywhere